+++ helpers/__init__.py
@@ -1,4 +1,4 @@
 # helpers package for lazyscan
 """Helper modules for lazyscan functionality"""
 
-__version__ = "0.3.1"
+__version__ = "0.5.0"